Check length/entropy of refresh token, authorization code

Closes #538

See merge request openid/conformance-suite!743

+++ src/main/java/io/fintechlabs/testframework/condition/client/EnsureMinimumAuthenticationRequestIdLength.java
@@ -7,7 +7,7 @@ import io.fintechlabs.testframework.testmodule.Environment;
 
 public class EnsureMinimumAuthenticationRequestIdLength extends AbstractCondition {
 
-	private final double requiredLength = 128;
+	private final int requiredLength = 128;
 
 	@Override
 	@PreEnvironment(required = "backchannel_authentication_endpoint_response")
